% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungage.R
\name{myrec_player_names}
\alias{myrec_player_names}
\title{Pull Player Names present within a recording}
\usage{
myrec_player_names(myrec_path)
}
\arguments{
\item{myrec_path}{Path to the myrec file}
}
\description{
Pull Player Names present within a recording
}
