% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{myrec_to_bids}
\alias{myrec_to_bids}
\title{Convert myrec file into BIDS format}
\usage{
myrec_to_bids(
  myrec_path,
  bids_dataset_object,
  session_id,
  task_id,
  participant_id_map
)
}
\arguments{
\item{participant_id_map}{Named vector with names as player names and values as participant IDs}
}
\description{
Convert myrec file into BIDS format
}
