% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungage.R
\name{myrec_to_bids}
\alias{myrec_to_bids}
\title{Convert myrec file into BIDS format}
\usage{
myrec_to_bids(
  myrec_path,
  bids_dataset_object,
  session_id,
  task_id,
  participant_id_map,
  person_sessions = 1000
)
}
\arguments{
\item{myrec_path}{Path to the .myrec file}

\item{bids_dataset_object}{BIDS object created by `rbids::bids()` to which this data will be written}

\item{session_id}{BIDS session id for this recording, one alphanumeric string}

\item{task_id}{BIDS task id for this recording, also an alphanumeric string}

\item{participant_id_map}{Named vector with names as player names and values as participant IDs}

\item{person_sessions}{Size of streams to read before writing data. Larger numbers run a little faster but use more memory. Default is 1000.}
}
\description{
Convert myrec file into BIDS format
}
